feat(improv): Extensions accept path string to denote required scope
Previously all extensions ran whenever improv rendered a template, now its possible to pass in a
specific path string that an extension will provide data for, so it's not run unless required. This
change also includes a fix to allow extensions access to the full middleware context.

This fixes #2, fixes  #3