feat(scene): Exit after enter callback if no path/branch added
Scene's without branches were leaving users infinitely engaged, unless manually exited because only
matching a branch will exit automatically. Now after the enter promise and callbacks are settled it
will check for existence of dialogue paths and exit the scene if none found. This required lots of
test and docs updates and a refactor of middleware.

Fixes #8. Breaks any code expecting scene.enter to resolve with a Dialogue. It will now resolve with
the full context and reject if middleware is interrupted.